ml/requirements.txt
absl-py==1.4.0
aio-pika==9.2.3
aiormq==6.7.7
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
boto3==1.28.49
botocore==1.31.49
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
jax==0.4.14
jmespath==1.0.1
keras==2.12.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pamqp==3.2.1
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-addons==0.21.0
termcolor==2.3.0
typeguard==2.13.3
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
Werkzeug==2.3.7
wrapt==1.14.1
yarl==1.9.2
zipp==3.16.2